Start of Cards to be Displayed/Comp Cards for Play
Created two functions for stripOutPlayer in order to have an array of what can be played/what is showen to the player to be played.
Also second function stripOutComputer which gives the array of which items the computer is able to play.
Need to create playerHand from the deltSuspects[0], deltWeapons[0], deltRooms[0] and compHand from deltSuspects[1], deltWeapons[1], deltRooms[1]